WebService.Tests created, Domain.Tests removed
