--- Signals/Domain.Tests/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="RhinoMocks" version="3.6.1" targetFramework="net461" />
-</packages>
