+++ Signals/Signals.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 14
-VisualStudioVersion = 14.0.24720.0
+VisualStudioVersion = 14.0.25420.1
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain", "Domain\Domain.csproj", "{7643117C-25CB-4B38-8D3D-EFA38E2DC719}"
 EndProject
@@ -19,7 +19,7 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bootstrapper", "Bootstrappe
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DatabaseMaintenanceTool", "DatabaseMaintenanceTool\DatabaseMaintenanceTool.csproj", "{0594060D-7ABD-47DA-8773-1C4E0E44399C}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Domain.Tests", "Domain.Tests\Domain.Tests.csproj", "{9BF523D9-B18D-4FD7-9523-F207E77B881C}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WebService.Tests", "WebService.Tests\WebService.Tests.csproj", "{1D359E41-E94D-4E70-9204-1CF8250ACE90}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,10 +59,10 @@ Global
 		{0594060D-7ABD-47DA-8773-1C4E0E44399C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{0594060D-7ABD-47DA-8773-1C4E0E44399C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{0594060D-7ABD-47DA-8773-1C4E0E44399C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9BF523D9-B18D-4FD7-9523-F207E77B881C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9BF523D9-B18D-4FD7-9523-F207E77B881C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9BF523D9-B18D-4FD7-9523-F207E77B881C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9BF523D9-B18D-4FD7-9523-F207E77B881C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1D359E41-E94D-4E70-9204-1CF8250ACE90}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1D359E41-E94D-4E70-9204-1CF8250ACE90}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1D359E41-E94D-4E70-9204-1CF8250ACE90}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1D359E41-E94D-4E70-9204-1CF8250ACE90}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
