+++ Signals/WebService.Tests/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Castle.Core" version="3.3.3" targetFramework="net461" />
+  <package id="Moq" version="4.5.10" targetFramework="net461" />
+</packages>
